test(sentry): fix test case to ensure that _schedule_downstream_tasks and _run_raw_task are decorated with error
